+++ cliquet/storage/__init__.py
@@ -58,8 +58,8 @@ class StorageBase(object):
         """
         try:
             auth = request.headers.get('Authorization')
-            if request.registry.settings.get('read_only'):
-                # Do not try to write in read_only mode.
+            if request.registry.settings.get('readonly'):
+                # Do not try to write in readonly mode.
                 self.get_all(_HEARTBEAT_COLLECTION_ID, _HEART_PARENT_ID,
                              auth=auth)
             else: